+++ appservice/src/index.ts
@@ -4,29 +4,32 @@
  *--------------------------------------------------------------------------------------------*/
 
 export * from './confirmOverwriteSettings';
-export { AppKind, LinuxRuntimes, WebsiteOS } from './createAppService/AppKind';
 export * from './createAppService/AppInsightsCreateStep';
 export * from './createAppService/AppInsightsListStep';
-export * from './createAppService/IAppServiceWizardContext';
-export * from './createAppService/AppServicePlanListStep';
+export { AppKind, LinuxRuntimes, WebsiteOS } from './createAppService/AppKind';
 export * from './createAppService/AppServicePlanCreateStep';
+export * from './createAppService/AppServicePlanListStep';
+export * from './createAppService/IAppServiceWizardContext';
+export * from './createAppService/setLocationsTask';
 export * from './createAppService/SiteCreateStep';
 export * from './createAppService/SiteHostingPlanStep';
 export * from './createAppService/SiteNameStep';
 export * from './createAppService/SiteOSStep';
 export * from './createAppService/SiteRuntimeStep';
-export * from './remoteDebug/remoteDebugCommon';
-export * from './remoteDebug/startRemoteDebug';
 export * from './createSlot';
+export * from './deleteSite';
 export * from './deploy/deploy';
 export * from './deploy/getDeployFsPath';
 export * from './deploy/runPreDeployTask';
-export * from './deleteSite';
 export * from './editScmType';
-export * from './getFile';
+export { registerAppServiceExtensionVariables } from './extensionVariables';
 export * from './functionsAdminRequest';
+export * from './getFile';
+export { IConnectToGitHubWizardContext } from './github/IConnectToGitHubWizardContext';
 export * from './pingFunctionApp';
 export * from './putFile';
+export * from './remoteDebug/remoteDebugCommon';
+export * from './remoteDebug/startRemoteDebug';
 export * from './SiteClient';
 export * from './startStreamingLogs';
 export * from './swapSlot';
@@ -36,5 +39,5 @@ export * from './tree/DeploymentsTreeItem';
 export * from './tree/DeploymentTreeItem';
 export * from './tree/ISiteTreeRoot';
 export * from './TunnelProxy';
-export { registerAppServiceExtensionVariables } from './extensionVariables';
-export { IConnectToGitHubWizardContext } from './github/IConnectToGitHubWizardContext';
+
+// Adding a comment here otherwise "source.organizeImports" will cause duplicate blank lines and tslint will complain
